RavenDB-11352 adderss the PR comments
still to-do:

- Collect the info who is connected to the centeral.
- Write tests to verify the correct behavior when we disable/enbale.
