+++ src/Raven.Client/Documents/Commands/GetTcpInfoForRemoteTaskCommand.cs
@@ -20,7 +20,7 @@ namespace Raven.Client.Documents.Commands
 
         public override HttpRequestMessage CreateRequest(JsonOperationContext ctx, ServerNode node, out string url)
         {
-            url = $"{node.Url}/databases/{node.Database}/info/remote-task/tcp?database={_remoteDatabase}&remote-task={_remoteTask}";
+            url = $"{node.Url}/info/remote-task/tcp?database={_remoteDatabase}&remote-task={_remoteTask}";
 
             RequestedNode = node;
             var request = new HttpRequestMessage
