+++ src/Raven.Server/Web/RequestHandler.cs
@@ -15,15 +15,11 @@ using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Primitives;
 using Raven.Client;
 using Raven.Client.Http;
-using Raven.Client.ServerWide;
 using Raven.Client.ServerWide.Commands;
 using Raven.Client.ServerWide.Operations.Certificates;
-using Raven.Server.Json;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide;
-using Raven.Server.ServerWide.Commands;
 using Raven.Server.ServerWide.Context;
-using Raven.Server.Web.System;
 using Sparrow;
 using Sparrow.Json;
 using Sparrow.Utils;
@@ -541,198 +537,5 @@ namespace Raven.Server.Web
         {
             SetupCORSHeaders(HttpContext);
         }
-
-        protected async Task PutCertificate(string prefix, bool includePrivateKey)
-        {
-            // one of the first admin action is to create a certificate, so let
-            // us also use that to indicate that we are the seed node
-            ServerStore.EnsureNotPassive();
-            using (ServerStore.ContextPool.AllocateOperationContext(out TransactionOperationContext ctx))
-            using (var certificateJson = ctx.ReadForDisk(RequestBodyStream(), "put-certificate"))
-            {
-                var certificate = JsonDeserializationServer.CertificateDefinition(certificateJson);
-                ValidateCertificateDefinition(certificate, ServerStore);
-
-                using (ctx.OpenReadTransaction())
-                {
-                    var clientCert = (HttpContext.Features.Get<IHttpAuthenticationFeature>() as RavenServer.AuthenticateConnection)?.Certificate;
-                    var clientCertDef = ReadCertificateFromCluster(ctx, prefix + clientCert?.Thumbprint);
-
-                    if ((certificate.SecurityClearance == SecurityClearance.ClusterAdmin || certificate.SecurityClearance == SecurityClearance.ClusterNode) && IsClusterAdmin() == false)
-                        throw new InvalidOperationException($"Cannot save the certificate '{certificate.Name}' with '{certificate.SecurityClearance}' security clearance because the current client certificate being used has a lower clearance: {clientCertDef.SecurityClearance}");
-
-                    if (string.IsNullOrWhiteSpace(certificate.Certificate))
-                        throw new ArgumentException($"{nameof(certificate.Certificate)} is a mandatory property when saving an existing certificate");
-                }
-
-                byte[] certBytes;
-                try
-                {
-                    certBytes = Convert.FromBase64String(certificate.Certificate);
-                }
-                catch (Exception e)
-                {
-                    throw new ArgumentException($"Unable to parse the {nameof(certificate.Certificate)} property, expected a Base64 value", e);
-                }
-
-                try
-                {
-                    var _ = string.IsNullOrEmpty(certificate.Password)
-                        ? new X509Certificate2(certBytes, (string)null, X509KeyStorageFlags.MachineKeySet)
-                        : new X509Certificate2(certBytes, certificate.Password, X509KeyStorageFlags.MachineKeySet);
-                }
-                catch (Exception e)
-                {
-                    throw new ArgumentException("Unable to load the provided certificate.", e);
-                }
-
-                try
-                {
-                    await PutCertificateCollectionInCluster(prefix, certificate, certBytes, certificate.Password, ServerStore, ctx, includePrivateKey);
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException($"Failed to put certificate {certificate.Name} in the cluster.", e);
-                }
-
-                NoContentStatus();
-                HttpContext.Response.StatusCode = (int)HttpStatusCode.Created;
-            }
-        }
-
-        public static async Task PutCertificateCollectionInCluster(
-            string prefix,
-            CertificateDefinition certDef,
-            byte[] certBytes, string password,
-            ServerStore serverStore,
-            TransactionOperationContext ctx,
-            bool includePrivateKey = false)
-        {
-            var collection = new X509Certificate2Collection();
-
-            if (string.IsNullOrEmpty(password))
-                collection.Import(certBytes, (string)null, X509KeyStorageFlags.MachineKeySet | X509KeyStorageFlags.Exportable);
-            else
-                collection.Import(certBytes, password, X509KeyStorageFlags.PersistKeySet | X509KeyStorageFlags.Exportable | X509KeyStorageFlags.MachineKeySet);
-
-            var first = true;
-            var collectionPrimaryKey = string.Empty;
-
-            // we don't want to import items up the chain (signing keys)
-            var issuers = new HashSet<string>();
-
-            foreach (var x509Certificate in collection)
-            {
-                if (serverStore.Server.Certificate.Certificate?.Thumbprint != null && serverStore.Server.Certificate.Certificate.Thumbprint.Equals(x509Certificate.Thumbprint))
-                    throw new InvalidOperationException($"You are trying to import the same server certificate ({x509Certificate.Thumbprint}) as the one which is already loaded. This is not supported.");
-
-                if (x509Certificate.Issuer != x509Certificate.Subject)
-                    issuers.Add(x509Certificate.Issuer);
-            }
-
-            foreach (var x509Certificate in collection)
-            {
-                if (issuers.Contains(x509Certificate.Subject))
-                    continue;
-
-                var currentCertDef = new CertificateDefinition
-                {
-                    Name = certDef.Name,
-                    Permissions = certDef.Permissions,
-                    SecurityClearance = certDef.SecurityClearance,
-                    Password = certDef.Password,
-                };
-
-                if (includePrivateKey)
-                {
-                    // here we need to keep the private key, since in pull replication we are the initiator of the communication.
-                    currentCertDef.Certificate = Convert.ToBase64String(x509Certificate.Export(X509ContentType.Pfx));
-                }
-                else
-                {
-                    // avoid storing the private key for client cert
-                    currentCertDef.Certificate = Convert.ToBase64String(x509Certificate.Export(X509ContentType.Cert));
-                }
-
-                // In case of a collection, we group all the certificates together and treat them as one unit. 
-                // They all have the same name and permissions but a different thumbprint.
-                // The first certificate in the collection will be the primary certificate and its thumbprint will be the one shown in a GET request
-                // The other certificates are secondary certificates and will contain a link to the primary certificate.
-                currentCertDef.Thumbprint = x509Certificate.Thumbprint;
-                currentCertDef.NotAfter = x509Certificate.NotAfter;
-
-                if (first)
-                {
-                    var firstKey = prefix + x509Certificate.Thumbprint;
-                    collectionPrimaryKey = firstKey;
-
-                    foreach (var cert in collection)
-                    {
-                        if (issuers.Contains(cert.Subject))
-                            continue;
-
-                        if (prefix + cert.Thumbprint != firstKey)
-                            currentCertDef.CollectionSecondaryKeys.Add(prefix + cert.Thumbprint);
-                    }
-                }
-                else
-                    currentCertDef.CollectionPrimaryKey = collectionPrimaryKey;
-
-
-                var certKey = prefix + currentCertDef.Thumbprint;
-                if (serverStore.CurrentRachisState == RachisState.Passive)
-                {
-                    using (var certificate = ctx.ReadObject(currentCertDef.ToJson(), "Client/Certificate/Definition"))
-                    using (var tx = ctx.OpenWriteTransaction())
-                    {
-                        serverStore.Cluster.PutLocalState(ctx, certKey, certificate);
-                        tx.Commit();
-                    }
-                }
-                else
-                {
-                    var putResult = await serverStore.PutValueInClusterAsync(new PutCertificateCommand(certKey, currentCertDef));
-                    await serverStore.Cluster.WaitForIndexNotification(putResult.Index);
-                }
-
-                first = false;
-            }
-        }
-
-        protected static void ValidateCertificateDefinition(CertificateDefinition certificate, ServerStore serverStore)
-        {
-            if (string.IsNullOrWhiteSpace(certificate.Name))
-                throw new ArgumentException($"{nameof(certificate.Name)} is a required field in the certificate definition");
-
-            ValidatePermissions(certificate, serverStore);
-        }
-
-        private static void ValidatePermissions(CertificateDefinition certificate, ServerStore serverStore)
-        {
-            if (certificate.Permissions == null)
-                throw new ArgumentException($"{nameof(certificate.Permissions)} is a required field in the certificate definition");
-
-            foreach (var kvp in certificate.Permissions)
-            {
-                if (string.IsNullOrWhiteSpace(kvp.Key))
-                    throw new ArgumentException("Error in permissions in the certificate definition, database name is empty");
-
-                if (ResourceNameValidator.IsValidResourceName(kvp.Key, serverStore.Configuration.Core.DataDirectory.FullPath, out var errorMessage) == false)
-                    throw new ArgumentException("Error in permissions in the certificate definition:" + errorMessage);
-
-                if (kvp.Value != DatabaseAccess.ReadWrite && kvp.Value != DatabaseAccess.Admin)
-                    throw new ArgumentException($"Error in permissions in the certificate definition, invalid access {kvp.Value} for database {kvp.Key}");
-            }
-        }
-
-        protected CertificateDefinition ReadCertificateFromCluster(TransactionOperationContext ctx, string key)
-        {
-            var certificate = ServerStore.Cluster.Read(ctx, key);
-            if (certificate == null)
-                return null;
-
-            return JsonDeserializationServer.CertificateDefinition(certificate);
-        }
-
     }
 }
